refactor(l1): unify discovery servers into single DiscoveryServer actor
