+++ crates/networking/p2p/discovery/mod.rs
@@ -1,7 +1,4 @@
-//! Discovery multiplexer for running both discv4 and discv5 on a shared UDP port.
-//!
-//! This module provides packet discrimination between discv4 and discv5 protocols
-//! and routes packets to the appropriate protocol handler.
+//! Discovery protocol implementation for running both discv4 and discv5 on a shared UDP port.
 //!
 //! ## Packet Discrimination Strategy
 //!
@@ -12,17 +9,37 @@
 //! **Discrimination logic:**
 //! 1. If packet length >= 98 bytes AND `packet[0..32] == keccak256(packet[32..])` → DiscV4
 //! 2. Otherwise → DiscV5
-//!
-//! This is O(1) with a single keccak hash and has negligible false positive probability (2^-256).
 
 pub mod codec;
-mod multiplexer;
+mod discv4_handlers;
+mod discv5_handlers;
+pub mod server;
 
-pub use multiplexer::{
-    DiscoveryConfig, DiscoveryMultiplexer, DiscoveryMultiplexerError, is_discv4_packet,
-};
+pub use server::{DiscoveryServer, DiscoveryServerError, is_discv4_packet};
 
 use std::time::Duration;
+
+/// Configuration for which discovery protocols to enable.
+#[derive(Debug, Clone)]
+pub struct DiscoveryConfig {
+    pub discv4_enabled: bool,
+    pub discv5_enabled: bool,
+    pub initial_lookup_interval: f64,
+}
+
+impl Default for DiscoveryConfig {
+    fn default() -> Self {
+        Self {
+            discv4_enabled: true,
+            discv5_enabled: true,
+            initial_lookup_interval: INITIAL_LOOKUP_INTERVAL_MS,
+        }
+    }
+}
+
+/// Lookup interval constants shared by discv4, discv5, and RLPx initiator.
+pub const INITIAL_LOOKUP_INTERVAL_MS: f64 = 100.0; // 10 per second
+pub const LOOKUP_INTERVAL_MS: f64 = 600.0; // 100 per minute
 
 /// Smooth easing curve for discovery lookup intervals based on peer completion progress.
 ///
@@ -36,7 +53,6 @@ pub fn lookup_interval_function(progress: f64, lower_limit: f64, upper_limit: f6
         1.0 - ((-2.0 * progress + 2.0).powf(3.0)) / 2.0
     };
     Duration::from_micros(
-        // Use `progress` here instead of `ease_in_out_cubic` for a linear function.
         (1000f64 * (ease_in_out_cubic * (upper_limit - lower_limit) + lower_limit)).round() as u64,
     )
 }
