+++ crates/networking/p2p/rlpx/initiator.rs
@@ -1,4 +1,4 @@
-use crate::discv4::server::{LOOKUP_INTERVAL_MS, lookup_interval_function};
+use crate::discovery::{LOOKUP_INTERVAL_MS, lookup_interval_function};
 use crate::peer_table::PeerTableServerProtocol as _;
 use crate::types::Node;
 use crate::{metrics::METRICS, network::P2PContext, rlpx::connection::server::PeerConnection};
